airline_reservation.java
Command-Line User Interface (CLI)-based Airline Reservation System provides a simple and easy way for checking seat availability and reserving a seat if it is available. 
The system will first present the user with the available options. View available seats, book a seat, and EXIT is available as a choice. 
The user must choose a useful option.
The operation will be carried out as per the input choice.