Library_management_system.java [using swing library]
This Library Management System allows you to add books details using a graphical user interface (GUI) application. 
This system is build using the Java Swing package which gives a flexible set of components that can be use to build a Graphical user interface(GUI). 
This application has an easy and simple design with buttons to add books, view books list, edit the details, delete the book, clear the text fields and exit application.