(event)CalendarApplication.java  (using swing library)
Event Calendar Application is a user-friendly desktop application built using Java Swing that allows users to manage their events and appointments efficiently. 
The application provides an interactive graphical interface for adding, editing, and deleting events, along with a clear display of scheduled events in a list. 
With its  design, this Application simplifies event management and scheduling for users.